+++ tclParse.c
@@ -1171,7 +1171,7 @@ Tcl_ParseVar(interp, string, termPtr)
 		    != TCL_OK) {
 		char msg[100];
 		sprintf(msg, "\n    (parsing index for array \"%.*s\")",
-			string-name1, name1);
+			(int)(string-name1), name1);
 		Tcl_AddErrorInfo(interp, msg);
 		result = NULL;
 		name2 = pv.buffer;
